Create Coffee is for Closers
Program built for MGMT 586. 
This program calculates the maximum loan principle amount and monthly payment for a car buyer, based on user inputs of annual gross income, credit score, and loan length. 

This program includes building and calling multi-pass functions, a while loop, returning function calculations to a variable outside of the function, and performing mathematical equations within a program.